% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstar.R
\name{field_value}
\alias{field_value}
\title{Accessor: a field's value by name.}
\usage{
field_value(ds, name)
}
\arguments{
\item{ds}{an \code{lstar_dataset}}

\item{name}{field name}
}
\value{
the field's \code{values} (a vector, matrix, or sparse \code{Matrix}), or \code{NULL} if absent.
}
\description{
Accessor: a field's value by name.
}
