% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstar.R
\name{print.lstar_dataset}
\alias{print.lstar_dataset}
\title{Print an L* dataset}
\usage{
\method{print}{lstar_dataset}(x, ...)
}
\arguments{
\item{x}{an \code{lstar_dataset}}

\item{...}{ignored, for S3 compatibility}
}
\value{
\code{x}, invisibly (called for the side effect of printing a summary of axes and fields).
}
\description{
Print an L* dataset
}
