% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_sce.R
\name{read_sce}
\alias{read_sce}
\title{Read a SingleCellExperiment into an L* dataset.}
\usage{
read_sce(sce)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment}}
}
\value{
an \code{lstar_dataset} of kind \code{"sample"}.
}
\description{
Read a SingleCellExperiment into an L* dataset.
}
\seealso{
\code{\link[=write_sce]{write_sce()}}
}
