% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_conos.R
\name{write_conos}
\alias{write_conos}
\title{Build an L* dataset from a Conos object (a collection of samples).}
\usage{
write_conos(co, clustering = NULL)
}
\arguments{
\item{co}{a \code{Conos} R6 object}

\item{clustering}{optional name of a joint clustering in \code{co$clusters} (default: all)}
}
\value{
an \code{lstar_dataset} of kind \code{collection}
}
\description{
Build an L* dataset from a Conos object (a collection of samples).
}
