% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_sce.R
\name{write_sce}
\alias{write_sce}
\title{Build a SingleCellExperiment from an L* dataset.}
\usage{
write_sce(ds)
}
\arguments{
\item{ds}{an \code{lstar_dataset}}
}
\value{
a \code{SingleCellExperiment} object.
}
\description{
Measures become assays (transposed to genes x cells), embeddings become \code{reducedDims}, and
arity-1 cell fields become \code{colData}.
}
\seealso{
\code{\link[=read_sce]{read_sce()}}
}
