% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_seurat.R
\name{write_seurat}
\alias{write_seurat}
\title{Build a Seurat object from an L* dataset.}
\usage{
write_seurat(ds)
}
\arguments{
\item{ds}{an \code{lstar_dataset}}
}
\value{
a \code{Seurat} object.
}
\description{
Measures over \verb{(cells, genes)} become assay layers (transposed to Seurat's genes x cells
orientation); embeddings become \code{DimReduc}s; arity-1 cell fields become \code{meta.data}.
}
\seealso{
\code{\link[=read_seurat]{read_seurat()}}
}
